requirements.txt
cloudpickle==3.1.1
contourpy==1.3.1
coverage==7.7.0
cycler==0.12.1
fonttools==4.56.0
imbalanced-learn==0.13.0
imblearn==0.0
iniconfig==2.1.0
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.1
numba==0.61.0
numpy==2.1.3
nvidia-nccl-cu12==2.26.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
pyparsing==3.2.1
pytest==8.3.5
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shap==0.47.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
xgboost==3.0.0
